[DFFRS-486] Fix dw_collection_recovery schema (field type) and NEGOCIAC (#12982)
* update schema - field type

* update sk_negotiation field type

* fix NEGOCIAC

* fix negociac

* fix dupicated records causes by customer name, and transform none values in agreement_in_delay into 'N' value

* deduplicated installment_canceled and update table description (snapshot start date)

* add creditor  name

+++ dags/fintech/dw_collection_recovery/queries/dw/fact_collection.sql
@@ -38,12 +38,12 @@ SELECT
     id_operator_registration,
     id_occurrence,
     occurrence,
-    SUM(esforco) AS total_esforco,
-    SUM(alo) AS total_alo,
-    SUM(cpc) AS total_cpc,
-    SUM(promisse) AS total_promisse,
-    SUM(agreement) AS total_agreement,
-    SUM(failure) AS total_failure,
+    INT(SUM(esforco)) AS total_esforco,
+    INT(SUM(alo)) AS total_alo,
+    INT(SUM(cpc)) AS total_cpc,
+    INT(SUM(promisse)) AS total_promisse,
+    INT(SUM(agreement)) AS total_agreement,
+    INT(SUM(failure)) AS total_failure,
     dt_occurrence,
     NOW() AS ts_load
 FROM collection_calculation
